Onboarding: Guard deployedURL
Before, it was possible to provide `https://foo.com/` as input and generate a redirect_uri like `https://foo.com//api/keystatic/github/oauth/callback` which triggers the "The `redirect_uri` is not associated with this application." error by Github. Now the path is merged with the URL correctly for any valid URL input.